// packages/wings/src/proto/wings/flight/fetch.ts
export interface FetchTicket {
  $type: "wings.flight.FetchTicket";
  readonly tableName: string;
  readonly partitionValue?: PartitionValue | undefined;
  readonly offset: bigint;
  readonly minBatchSize: number;
  readonly maxBatchSize: number;
}

function createBaseFetchTicket(): FetchTicket {
  return {
    $type: "wings.flight.FetchTicket",
    tableName: "",
    partitionValue: undefined,
    offset: 0n,
    minBatchSize: 0,
    maxBatchSize: 0,
  };
}

export const FetchTicket: MessageFns<FetchTicket, "wings.flight.FetchTicket"> = {
  $type: "wings.flight.FetchTicket" as const,

  encode(message: FetchTicket, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.tableName !== "") {
      writer.uint32(10).string(message.tableName);
    }
    if (message.partitionValue !== undefined) {
      PartitionValue.encode(message.partitionValue, writer.uint32(18).fork()).join();
    }
    if (message.offset !== 0n) {
      if (BigInt.asUintN(64, message.offset) !== message.offset) {
        throw new globalThis.Error(
          "value provided for field message.offset of type uint64 too large",
        );
      }
      writer.uint32(24).uint64(message.offset);
    }
    if (message.minBatchSize !== 0) {
      writer.uint32(32).uint32(message.minBatchSize);
    }
    if (message.maxBatchSize !== 0) {
      writer.uint32(40).uint32(message.maxBatchSize);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FetchTicket {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFetchTicket() as any;
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tableName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.partitionValue = PartitionValue.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.offset = reader.uint64() as bigint;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.minBatchSize = reader.uint32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.maxBatchSize = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FetchTicket {
    return {
      $type: FetchTicket.$type,
      tableName: isSet(object.tableName) ? globalThis.String(object.tableName) : "",
      partitionValue: isSet(object.partitionValue)
        ? PartitionValue.fromJSON(object.partitionValue)
        : undefined,
      offset: isSet(object.offset) ? BigInt(object.offset) : 0n,
      minBatchSize: isSet(object.minBatchSize) ? globalThis.Number(object.minBatchSize) : 0,
      maxBatchSize: isSet(object.maxBatchSize) ? globalThis.Number(object.maxBatchSize) : 0,
    };
  },

  toJSON(message: FetchTicket): unknown {
    const obj: any = {};
    if (message.tableName !== "") {
      obj.tableName = message.tableName;
    }
    if (message.partitionValue !== undefined) {
      obj.partitionValue = PartitionValue.toJSON(message.partitionValue);
    }
    if (message.offset !== 0n) {
      obj.offset = message.offset.toString();
    }
    if (message.minBatchSize !== 0) {
      obj.minBatchSize = Math.round(message.minBatchSize);
    }
    if (message.maxBatchSize !== 0) {
      obj.maxBatchSize = Math.round(message.maxBatchSize);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FetchTicket>, I>>(base?: I): FetchTicket {
    return FetchTicket.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FetchTicket>, I>>(object: I): FetchTicket {
    const message = createBaseFetchTicket() as any;
    message.tableName = object.tableName ?? "";
    message.partitionValue =
      object.partitionValue !== undefined && object.partitionValue !== null
        ? PartitionValue.fromPartial(object.partitionValue)
        : undefined;
    message.offset = object.offset ?? 0n;
    message.minBatchSize = object.minBatchSize ?? 0;
    message.maxBatchSize = object.maxBatchSize ?? 0;
    return message;
  },
};

messageTypeRegistry.set(FetchTicket.$type, FetchTicket);

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends { readonly $case: string }
        ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { readonly $case: T["$case"] }
        : T extends {}
          ? { [K in Exclude<keyof T, "$type">]?: DeepPartial<T[K]> }
          : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P> | "$type">]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T, V extends string> {
  readonly $type: V;
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}